moved benchmark for threshold sauvola in benchmark_filters .py and removed the benchmark_threshold_sauvola.py file.
